requirements.txt
pyreadstat>=1.2.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
flask>=3.0.0
openpyxl>=3.1.0
gunicorn>=22.0.0
pytest>=8.0.0
